pyproject.toml
[tool.poetry]
name = "lebedoo-api"
version = "0.1.0"
description = ""
authors = ["v015h3bn1k <jacky.rebeiz@gmail.com>"]
readme = "README.md"
packages = [{include = "lebedoo_api"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.1.1"
graphene-django = "^3.0.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
psycopg2 = "^2.9.3"
django-graphql-jwt = "^0.3.4"
psycopg2-binary = "^2.9.4"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = {version = "^22.10.0", allow-prereleases = true}
mypy = "^0.991"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
